Deprecated use of graph_detail::is_same. Removed usage from visitors header, and tagged the defining file with the #warning.
[SVN r53151]

+++ test/tiernan_all_cycles.cpp
@@ -52,7 +52,7 @@ void test()
     // Generate random graphs with 15 vertices and 15% probability
     // of edge connection.
     static const size_t N = 20;
-    static const float P = 0.1;
+    static const double P = 0.1;
     boost::minstd_rand rng;
 
     Graph g(er(rng, N, P), er(), N);
